requirements.txt
# 核心依赖
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-dotenv>=0.19.0

# 数据处理
pyarrow>=10.0.0
openpyxl>=3.0.0

# 机器学习
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0
optuna>=3.0.0

# 深度学习
torch>=2.0.0
pytorch-lightning>=2.0.0

# API框架
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0

# 工具库
pyyaml>=6.0
click>=8.0.0
tqdm>=4.65.0
joblib>=1.2.0
chinese-calendar>=1.8.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0

# 测试
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# 监控和日志
prometheus-client>=0.16.0
python-json-logger>=2.0.0